Libs/Reflex/src/main/java/reflex/util/ValueSerializer.java
/**
 * The MIT License (MIT)
 *
 * Copyright (c) 2011-2016 Incapture Technologies LLC
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
package reflex.util;

import rapture.common.impl.jackson.JacksonUtil;
import reflex.node.KernelExecutor;
import reflex.value.ReflexValue;

/**
 * A good serializer of ReflexValues
 * @author amkimian
 *
 */
public class ValueSerializer {
    public static String serialize(ReflexValue value) {
        return KernelExecutor.convertValueToJson(value,true);
    }
    
    public static ReflexValue deserialize(String json) throws ClassNotFoundException {
        Object x = JacksonUtil.objectFromJson(json, Object.class);
        return KernelExecutor.reconstructFromObject(x);
    }
}
